requirements.txt
pip==22.3.1
scikit-learn==1.2.0
opencv-python
h5py
matplotlib
Pillow
requests
psutil
imutils
pandas
streamlit
streamlit_option_menu
keras==2.11.0
tensorflow==2.11.0